engine: start work on esm/utilities/dialog.js
